--- flutter/ottugi_curry/lib/view/page/main/main_3d_page.dart
@@ -1,68 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:model_viewer_plus/model_viewer_plus.dart';
-import 'package:get/get.dart';
-
-/// 3D 라이브러리 테스트 중인 임시 화면
-class Main3DPage extends StatelessWidget {
-  const Main3DPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        leading: Padding(
-          padding: const EdgeInsets.only(left: 25),
-          child: IconButton(
-            icon: const Icon(
-              Icons.arrow_back_ios,
-              color: Colors.black,
-            ),
-            onPressed: () {
-              Get.back();
-            },
-          ),
-        ),
-      ),
-      body: ModelViewer(
-        backgroundColor: Color.fromARGB(0xFF, 0xEE, 0xEE, 0xEE),
-        src: 'assets/3d_models/Astronaut.glb',
-        alt: 'A 3D model of an astronaut',
-        ar: true,
-        arModes: ['scene-viewer', 'webxr', 'quick-look'],
-        // autoRotate: true,
-        iosSrc: 'https://modelviewer.dev/shared-assets/models/Astronaut.usdz',
-        disableZoom: true,
-      ),
-    );
-  }
-}
-
-class Main2DPage extends StatelessWidget {
-  const Main2DPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(),
-      body: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Text('방을 360도로 구경하려면 이미지를 눌러주세요!'),
-          InkWell(
-            onTap: () {
-              Get.to(() => Main3DPage());
-            },
-            child: SizedBox(
-              height: 500,
-              child: Image.asset(
-                'assets/3d_models/Astronaut_2d.png',
-                fit: BoxFit.fitHeight,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
